pyproject.toml
[project]
name = "jogo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame>=2.6.1",
    "pygame-ce>=2.5.7",
    "pyopengl>=3.1.10",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
